+++ fragments/labels/dialog.sh
@@ -1,8 +1,8 @@
 dialog)
     name="Dialog"
     type="pkg"
-    downloadURL=$(downloadURLFromGit bartreardon Dialog-public )
-    appNewVersion=$(versionFromGit bartreardon Dialog-public )
+    packageID="au.csiro.dialogcli"
+    downloadURL="$(downloadURLFromGit bartreardon Dialog)"
+    appNewVersion="$(versionFromGit bartreardon Dialog)"
     expectedTeamID="PWA5E9TQ59"
     ;;
-